requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
google-genai
faiss-cpu
numpy